@@ -1,6 +1,6 @@
 {
   "name": "@fluentui/react-icons-font-subsetting-webpack-plugin",
-  "version": "2.0.310",
+  "version": "2.0.315",
   "description": "Webpack plugin to subset the icon fonts used by @fluentui/react-icons based on which icons are used.",
   "main": "lib/index.js",
   "scripts": {
@@ -22,7 +22,7 @@
   },
   "homepage": "https://github.com/microsoft/fluentui-system-icons/packages/react-icons-font-subsetting-webpack-plugin#readme",
   "devDependencies": {
-    "typescript": "^4.6.3",
+    "typescript": "5.0.4",
     "webpack": "^5.72.0",
     "@fluentui/react-icons": "*",
     "@types/node": "14"
