@@ -12,7 +12,16 @@ module.exports = {
     },
     resolve: {
         // Add `.ts` and `.tsx` as a resolvable extension.
-        extensions: [".ts", ".tsx", ".js"]
+        extensions: [".ts", ".tsx", ".js"],
+        alias: {
+            // Force React and React-DOM to use the versions installed in icon-app (React 19)
+            // This prevents version conflicts with react-icons package which uses React 17
+            'react': path.resolve(__dirname, 'node_modules/react'),
+            'react-dom': path.resolve(__dirname, 'node_modules/react-dom'),
+            // Ensure React/jsx-runtime uses the correct version
+            'react/jsx-runtime': path.resolve(__dirname, 'node_modules/react/jsx-runtime'),
+            'react/jsx-dev-runtime': path.resolve(__dirname, 'node_modules/react/jsx-dev-runtime')
+        }
     },
     mode: 'development',
     module: {
@@ -29,5 +38,5 @@ module.exports = {
         new HtmlWebpackPlugin({
             title: 'Fluent React Icons Demo'
         })
-    ]
+    ],
 };
