@@ -1,10 +1,10 @@
 import * as React from "react";
-import { FluentIconsProps } from "./FluentIconsProps.types";
+import type { FluentIconsProps } from "./FluentIconsProps.types";
 import { useIconState } from "./useIconState";
-import { CreateFluentIconOptions, FluentIcon } from "./createFluentIcon";
+import type { CreateFluentIconOptions, FluentIcon } from "./createFluentIcon";
 
-const wrapIcon = (Icon: (iconProps: FluentIconsProps) => JSX.Element, displayName?: string, options?: CreateFluentIconOptions) => {
-    const WrappedIcon = React.forwardRef((props: FluentIconsProps, ref: FluentIconsProps["ref"]) => { 
+const wrapIcon = (Icon: (iconProps: FluentIconsProps) => React.ReactElement, displayName?: string, options?: CreateFluentIconOptions) => {
+    const WrappedIcon = React.forwardRef((props: FluentIconsProps, ref: FluentIconsProps["ref"]) => {
         const state = {
             ...useIconState(props, { flipInRtl: options?.flipInRtl }),
             ref
@@ -14,6 +14,6 @@ const wrapIcon = (Icon: (iconProps: FluentIconsProps) => JSX.Element, displayNam
     WrappedIcon.displayName = displayName;
     return WrappedIcon;
 }
- 
+
 
 export default wrapIcon;
