@@ -1,5 +1,23 @@
 import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
-  test: {},
+  test: {
+    globals: true,
+    setupFiles: ['./test-setup.js'],
+  },
+  plugins: [
+    {
+      name: 'mock-fonts',
+      resolveId(id) {
+        if (id.match(/\.(ttf|woff|woff2)$/)) {
+          return id;
+        }
+      },
+      load(id) {
+        if (id.match(/\.(ttf|woff|woff2)$/)) {
+          return 'export default "test-font-file"';
+        }
+      },
+    },
+  ],
 })
